ascii-animation-player/requirements.txt
opencv-python>=4.8.0
Pillow>=10.0.0
ascii-magic>=2.3.0
numpy>=1.26.0
colorama>=0.4.6
tqdm>=4.66.0
click>=8.1.0
pyyaml>=6.0 